+++ src/components/NewsCards/NewsCards.jsx
@@ -2,13 +2,7 @@ import { Grid, Grow, Typography } from '@mui/material';
 import React from 'react'
 import NewsCard from '../NewsCard/NewsCard';
 import useStyles from './styles';
-
-const infoCards = [
-    { color: '#00838f', title: 'Latest News', text: 'Give me the latest news' },
-    { color: '#1565c0', title: 'News by Categories/Countries', info: 'Business, Entertainment, General, Health, Science, Sports, Technology (from <Your_Country>)', text: 'Give me the latest Technology news from United States' },
-    { color: '#4527a0', title: 'News by Terms', info: 'Bitcoin, PlayStation 5, Smartphones, Donald Trump...', text: 'What\'s up with PlayStation 5' },
-    { color: '#283593', title: 'News by Sources', info: 'CNN, Wired, BBC News, Time, IGN, Buzzfeed, ABC News...', text: 'Give me the news from CNN' },
-];
+import infoCards from '../data/infoCards'
 
 const NewsCards = ({articles, activeArticle}) => {
